% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecutePhenotypeLibraryDiagnostics.R
\name{executePhenotyeLibraryDiagnostics}
\alias{executePhenotyeLibraryDiagnostics}
\title{Execute the cohort diagnostics}
\usage{
executePhenotyeLibraryDiagnostics(
  connectionDetails,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  verifyDependencies = TRUE,
  outputFolder,
  cohortIds = NULL,
  incrementalFolder = file.path(outputFolder, "incrementalFolder"),
  databaseId = "Unknown",
  databaseName = databaseId,
  databaseDescription = databaseId,
  extraLog = NULL
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created
using the
\code{\link[DatabaseConnector]{createConnectionDetails}}
function in the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM
format resides. Note that for SQL Server, this should
include both the database and schema name, for example
'cdm_data.dbo'.}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary data resides. This
is commonly the same as cdmDatabaseSchema. Note that for
SQL Server, this should include both the database and
schema name, for example 'vocabulary.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You
will need to have write privileges in this schema. Note
that for SQL Server, this should include both the
database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work
database schema. This table will hold the exposure and
outcome cohorts used in this study.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not
truly support temp tables. To emulate temp tables,
provide a schema with write privileges where temp tables
can be created.}

\item{verifyDependencies}{Check whether correct package versions are installed?}

\item{outputFolder}{Name of local folder to place results; make sure to use
forward slashes (/). Do not use a folder on a network
drive since this greatly impacts performance.}

\item{cohortIds}{An optional parameter to filter the cohortIds in OHDSI Phenotype library to run.
By Default all cohorts will be run.}

\item{incrementalFolder}{Name of local folder to hold the logs for incremental
run; make sure to use forward slashes (/). Do not use a
folder on a network drive since this greatly impacts
performance.}

\item{databaseId}{A short string for identifying the database (e.g.
'Synpuf').}

\item{databaseName}{The full name of the database (e.g. 'Medicare Claims
Synthetic Public Use Files (SynPUFs)').}

\item{databaseDescription}{A short description (several sentences) of the database.}

\item{extraLog}{Do you want to add anything extra into the log?}
}
\description{
Execute the cohort diagnostics
}
\details{
This function executes the cohort diagnostics.
}
