% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeLibraryDiagnostics.R
\docType{package}
\name{PhenotypeLibraryDiagnostics-package}
\alias{PhenotypeLibraryDiagnostics}
\alias{PhenotypeLibraryDiagnostics-package}
\title{PhenotypeLibraryDiagnostics: Generating Cohort Diagnostics for the OHDSI Phenotype Library}
\description{
Generating cohort diagnostics for the cohort definitions in the OHDSI Phenotype Library.
}
\keyword{internal}
